requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
structlog
crewai
langgraph
langchain-google-genai
google-generativeai
pyautogen<0.3
python-dotenv
